requirements.txt
fastmcp>=2.0.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
httpx>=0.27.0
pydantic>=2.0.0